finished project for geometry intersection
Add the function as intersection of line-circle, circle-circle.
Fix 3 bug which can influence the accuracy, but they are too small to report.
Update the hash function to get improvement of effecience.

+++ src/IO_class.cpp
@@ -41,7 +41,7 @@ vector<string> *IO_class::read()
 	return &all_geometry_string;
 }
 
-void IO_class::write(int n)
+void IO_class::write(size_t n)
 {
 	out << n << endl;
 }
